Avoid copying vector to Python bytes. Refactor summary parsing from file in a separate method.
