+++ eclair/src/main.rs
@@ -1,6 +1,4 @@
-use eclair::eclipse_binary::BinFile;
 use eclair::eclipse_summary::Summary;
-use eclair::errors::FileError;
 
 use anyhow as ah;
 use env_logger::{Builder, Env};
@@ -46,38 +44,23 @@ fn init_logger() {
 }
 
 fn main() -> ah::Result<()> {
-    // Initialize the logger as soon as possible
+    // Initialize the logger as soon as possible.
     init_logger();
 
-    // Read the CLI arguments
+    // Read the CLI arguments.
     let opt = Opt::from_args();
-
     let input_path = opt.input;
 
-    // If there is no stem, bail early
-    if input_path.file_stem().is_none() {
-        return Err(FileError::InvalidFilePath.into());
-    }
+    // Parse SMSPEC & UNSMRY files.
+    let summary = Summary::from_path(&input_path)?;
 
-    // we allow either extension or no extension at all
-    if let Some(ext) = input_path.extension() {
-        let ext = ext.to_str();
-        if ext != Some("SMSPEC") && ext != Some("UNSMRY") {
-            return Err(FileError::InvalidFileExt.into());
-        }
-    }
-
-    let smspec = BinFile::new(input_path.with_extension("SMSPEC"))?;
-    let unsmry = BinFile::new(input_path.with_extension("UNSMRY"))?;
-
-    let summary = Summary::new(smspec, unsmry)?;
-
+    // Create an output file.
     let mut out_file = File::create(
         opt.output
             .unwrap_or_else(|| input_path.with_extension("mpk")),
     )?;
 
-    // serialize summary data in the MessagePack format
+    // serialize summary data to the output file in the MessagePack format.
     let mut se = rmps::encode::Serializer::new(&mut out_file)
         .with_struct_map()
         .with_string_variants();
