+++ eclpy/src/lib.rs
@@ -1,6 +1,4 @@
-use eclair::eclipse_binary::BinFile;
 use eclair::eclipse_summary::Summary;
-use eclair::errors::FileError;
 
 use pyo3::exceptions;
 use pyo3::prelude::*;
@@ -9,50 +7,34 @@ use pyo3::wrap_pyfunction;
 use rmp_serde as rmps;
 use serde::Serialize;
 
-use std::path::Path;
+use std::{mem, path::Path};
 
 #[pyfunction]
 /// Load Eclipse's summary data in the MessegaPack format.
-fn as_msgpack(input_path: &str) -> PyResult<PyObject> {
+fn as_msgpack<'a: 'p, 'p>(py: Python<'p>, input_path: &'a str) -> PyResult<&'p PyBytes> {
     let input_path = Path::new(input_path);
-    // If there is no stem, bail early
-    if input_path.file_stem().is_none() {
-        return Err(exceptions::ValueError::py_err(
-            FileError::InvalidFilePath.to_string(),
-        ));
-    }
 
-    // we allow either extension or no extension at all
-    if let Some(ext) = input_path.extension() {
-        let ext = ext.to_str();
-        if ext != Some("SMSPEC") && ext != Some("UNSMRY") {
-            return Err(exceptions::ValueError::py_err(
-                FileError::InvalidFileExt.to_string(),
-            ));
-        }
-    }
-
-    let smspec = BinFile::new(input_path.with_extension("SMSPEC"))
-        .map_err(|err| exceptions::IOError::py_err(err.to_string()))?;
-    let unsmry = BinFile::new(input_path.with_extension("UNSMRY"))
+    // Parse SMSPEC & UNSMRY files.
+    let summary = Summary::from_path(input_path)
         .map_err(|err| exceptions::IOError::py_err(err.to_string()))?;
 
-    let summary =
-        Summary::new(smspec, unsmry).map_err(|err| exceptions::IOError::py_err(err.to_string()))?;
-
-    // serialize summary data in the MessagePack format
-    let mut wr = Vec::with_capacity(128);
-    let mut se = rmps::encode::Serializer::new(&mut wr)
+    // Serialize summary data to vector in the MessagePack format.
+    let mut vec = Vec::with_capacity(128);
+    let mut se = rmps::encode::Serializer::new(&mut vec)
         .with_struct_map()
         .with_string_variants();
+
     summary
         .serialize(&mut se)
         .map_err(|err| exceptions::IOError::py_err(err.to_string()))?;
 
-    let gil = Python::acquire_gil();
-    let py = gil.python();
-    let b = PyBytes::new(py, &wr);
-    Ok(b.to_object(py))
+    // Make Python bytes object from the vector. Forget it so that Rust does not clear its contents.
+    let ptr = vec.as_ptr();
+    let len = vec.len();
+    mem::forget(vec);
+
+    let b = unsafe { PyBytes::from_ptr(py, ptr, len) };
+    Ok(b)
 }
 
 #[pymodule]
